requirements.txt
asgiref==3.9.1
attrs==25.3.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==45.0.6
dj-rest-auth==7.0.1
Django==5.2.5
django-allauth==65.11.0
django-cors-headers==4.7.0
django-js-asset==3.1.2
django-mptt==0.18.0
djangorestframework==3.16.1
drf-spectacular==0.28.0
drf-yasg==1.21.10
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jwt==1.4.0
MarkupSafe==3.0.2
packaging==25.0
pillow==11.3.0
psycopg==3.3.0
psycopg-binary==3.3.0
pycparser==2.22
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
